requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.18
asyncpg==0.27.0
psycopg2-binary==2.9.6
pydantic==1.10.9
alembic==1.10.0
validators==0.18.2
httpx==0.24.1
greenlet
pytest-asyncio==0.24.0
pytest-mock
